Set unique, accurate titles for each authentication page
